refactor(tui): extract app-server skills routing from app
Move the temporary app-server `skills/list` request and response
adaptation helpers out of `tui/src/app.rs` into
`tui/src/app/skills.rs`.

This keeps the main app loop focused on orchestration, makes the
hybrid migration code easier to isolate, and removes a stale
`ListSkills`-specific comment from the `use_app_server` flag docs.
